Bug fixes. SOME CUSTOM SHORTCUTS WILL BE GONE!
fixed the weighted normals operator not enabling autosmooth. Updated the purge operator to work with multiple objects. Updated all of the menu operators to work with blender's naming conventions, to get rid of the terminal errors (functionality didn't change). WARNING: due to the nature of the update, any previously set MENU custom shortcuts will be RESET